+++ MetaCocktailsSwiftData/Views/Old Search View/Sub List Views/BasicComponentSearchListView.swift
